refactor sqlite reaction error handling
